Legato debugging
1. Add a bunch of printouts for tracing note status.
2. Completely disable the legato dual voice paradigm, so that only a single
   voice is used in legato mode. Only done fully for ADsynth so far, by
   simplifying the state machine updating the voice parameters.

First impression is that this works just as well as the previous incarnation
with two voices. However, there is something fishy here. It looks like part
of the whole point of having two voices is so that the voice parameters
can be calculated for the target pitch rather than the existing pitch, to
avoid waveform artefacts when the waveform parameters for the target pitch
don't match the source pitch. However, even with the dual voice paradigm,
there is tonal shift when playing legato, more noticable over long distances
on the keyboard. My conclusion therefore is that this feature is broken, and
should either be disabled, or perhaps better, fixed.

It could also be that it is only broken for certain synth engines.

More investigation is needed.